# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProductStatus = typeof ProductStatus[keyof typeof ProductStatus];


export const ProductStatus = {
  pending: 'pending',
  published: 'published',
  paused: 'paused',
  rejected: 'rejected',
} as const;

export interface Product {
  id: string;
  slug: string;
  title: string;
  shortDescription: string;
  description: string;
  imageUrl: string;
  category: string;
  ycBatch: string;
  websiteUrl: string;
  location: string;
  creatorName: string;
  creatorId: string;
  voteCount: number;
  ratingAverage: number;
  totalRaised: number;
  status: ProductStatus;
  featured: boolean;
  createdAt: string;
}

export interface ProductInput {
  /** @minLength 3 */
  title: string;
  /** @minLength 10 */
  shortDescription: string;
  /** @minLength 20 */
  description: string;
  imageUrl: string;
  /** @minLength 2 */
  category: string;
  /** @minLength 2 */
  ycBatch: string;
  websiteUrl: string;
  /** @minLength 2 */
  location: string;
}

export type StatusUpdateStatus = typeof StatusUpdateStatus[keyof typeof StatusUpdateStatus];


export const StatusUpdateStatus = {
  pending: 'pending',
  paused: 'paused',
} as const;

export interface StatusUpdate {
  status: StatusUpdateStatus;
}

export interface Ranking {
  rank: number;
  product: Product;
}

export interface CreatorDashboard {
  totalVotes: number;
  totalRaised: number;
  publishedCount: number;
  pendingCount: number;
  products: Product[];
}

export interface CheckoutInput {
  productId: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  disclosureAccepted: boolean;
}

export interface CheckoutSession {
  paymentId: string;
  checkoutUrl: string;
  amount: number;
  disclosure: string;
}

export type TransactionStatus = typeof TransactionStatus[keyof typeof TransactionStatus];


export const TransactionStatus = {
  pending: 'pending',
  paid: 'paid',
  failed: 'failed',
  refunded: 'refunded',
} as const;

export interface Transaction {
  id: string;
  productId: string;
  productTitle: string;
  amount: number;
  /**
     * @minimum 1
     * @maximum 5
     * @nullable
     */
  rating: number | null;
  status: TransactionStatus;
  disclosure: string;
  /** @nullable */
  receiptUrl?: string | null;
  createdAt: string;
}

export type ListProductsParams = {
search?: string;
category?: string;
sort?: ListProductsSort;
};

export type ListProductsSort = typeof ListProductsSort[keyof typeof ListProductsSort];


export const ListProductsSort = {
  trending: 'trending',
  newest: 'newest',
  community: 'community',
} as const;