# lib/api-zod/src/generated/types/checkoutInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CheckoutInput {
  productId: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  disclosureAccepted: boolean;
}