# lib/api-zod/src/generated/types/creatorDashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from './product';

export interface CreatorDashboard {
  totalVotes: number;
  totalRaised: number;
  publishedCount: number;
  pendingCount: number;
  products: Product[];
}