# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductStatus } from './productStatus';

export interface Product {
  id: string;
  slug: string;
  title: string;
  shortDescription: string;
  description: string;
  imageUrl: string;
  category: string;
  ycBatch: string;
  websiteUrl: string;
  location: string;
  creatorName: string;
  creatorId: string;
  voteCount: number;
  ratingAverage: number;
  totalRaised: number;
  status: ProductStatus;
  featured: boolean;
  createdAt: Date;
}