# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  /** @minLength 3 */
  title: string;
  /** @minLength 10 */
  shortDescription: string;
  /** @minLength 20 */
  description: string;
  imageUrl: string;
  /** @minLength 2 */
  category: string;
  /** @minLength 2 */
  ycBatch: string;
  websiteUrl: string;
  /** @minLength 2 */
  location: string;
}