# lib/api-zod/src/generated/types/ranking.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from './product';

export interface Ranking {
  rank: number;
  product: Product;
}