# lib/api-zod/src/generated/types/transaction.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionStatus } from './transactionStatus';

export interface Transaction {
  id: string;
  productId: string;
  productTitle: string;
  amount: number;
  /**
     * @minimum 1
     * @maximum 5
     * @nullable
     */
  rating: number | null;
  status: TransactionStatus;
  disclosure: string;
  /** @nullable */
  receiptUrl?: string | null;
  createdAt: Date;
}